# gala-ragdoll/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "A"
  version: "1.0"
  title: "Configuration traceability"
...
paths:
...
  /host/addHost:
    post:
      tags:
      - "host"
      summary: "add host in the configuration domain"
      description: "add host in the configuration domain"
      operationId: "add_host_in_domain"
      parameters:
      - in: "body"
        name: "body"
        description: "domain info"
        required: false
        schema:
          $ref: "#/definitions/HostInfos"
      responses:
        "200":
          description: "add host successfully in domain"
          schema:
            $ref: "#/definitions/BaseResponse"
      x-swagger-router-controller: "swagger_server.controllers.host_controller"
  /host/getHost:
    get:
      tags:
      - "host"
      summary: "get host by domainName"
      description: "get the host information of the configuration domain"
      operationId: "get_host_by_domain_name"
      parameters:
      - in: "body"
        name: "body"
        description: "domain info"
        required: false
        schema:
          $ref: "#/definitions/DomainName"
      responses:
        "200":
          description: "get host successfully in domain"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Host"
      x-swagger-router-controller: "swagger_server.controllers.host_controller"
  /host/deleteHost:
    delete:
      tags:
      - "host"
      summary: "delete host in the configuration  domain"
      description: "delete the host in the configuration domain"
      operationId: "delete_host_in_domain"
      parameters:
      - in: "body"
        name: "body"
        description: "domain info"
        required: false
        schema:
          $ref: "#/definitions/HostInfos"
      responses:
        "200":
          description: "delete host successfully in domain"
          schema:
            $ref: "#/definitions/BaseResponse"
      x-swagger-router-controller: "swagger_server.controllers.host_controller"
...
definitions:
...
  DomainName:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
  HostInfos:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      hostInfos:
        type: "array"
        items:
          $ref: "#/definitions/Host"
  Host:
    type: "object"
    required:
    - "hostId"
    properties:
      hostId:
        type: "string"
        description: "the id of host"
      ip:
        type: "string"
        description: "the ipv4 address of host"
      ipv6:
        type: "string"
        description: "the ipv6 address of host"
    example:
      ipv6: "ipv6"
      ip: "ip"
      hostId: "hostId"
...
  BaseResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
        description: "the response of the restfulApi"
    example:
      msg: "msg"
      code: 0
...